moved relevant code to schema/iframe
